auxiliary classes for distributed_vector separated into segments.hpp file
